--- a/lib/gwdb-legacy/dune
+++ b/lib/gwdb-legacy/dune
@@ -10,6 +10,7 @@
   )
   (modules_without_implementation dbdisk)
   (modules
+    btree
     database
     dbdisk
     dutil