*Luckily*, the internal data representation is the same, so we can still use old marshaled trees, which are compatible with new ones.

`Map.update` function which is way faster than `find`+`add`. On a 2 351 652 individuals database, I obtain the following speed up:

surname index creation time : ~28s -> ~16s
first name index creation time : ~25s -> ~14s